# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/juegos/{id}": {
      "get": {
        "operationId": "JuegoController_buscarPorId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Juego"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver el detalle de un juego (público)",
        "tags": [
          "Juegos"
        ]
      },
      "patch": {
        "operationId": "JuegoController_actualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarJuegoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Juego"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: el dato ya existe (duplicado) o está en uso y no se puede borrar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modificar un juego; las listas de ids enviadas reemplazan a las anteriores (solo ADMIN)",
        "tags": [
          "Juegos"
        ]
      },
      "delete": {
        "operationId": "JuegoController_eliminar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un juego (solo ADMIN)",
        "tags": [
          "Juegos"
        ]
      }
    },
...
    "/colecciones/{id}/juegos/{juegoId}": {
      "delete": {
        "operationId": "ColeccionController_quitarJuego",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "juegoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "CUU Administrar colección: quitar un juego (el dueño o un ADMIN)",
        "tags": [
          "Colecciones"
        ]
      }
    },
    "/biblioteca": {
      "get": {
        "operationId": "JuegoGuardadoController_listar",
        "parameters": [
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ME_INTERESA",
                "YA_JUGADO"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JuegoGuardado"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver mi biblioteca, con filtro opcional por estado (usuario con sesión)",
        "tags": [
          "Biblioteca"
        ]
      },
      "post": {
        "operationId": "JuegoGuardadoController_guardar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardarJuegoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JuegoGuardado"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: el dato ya existe (duplicado) o está en uso y no se puede borrar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Guardar un juego en mi biblioteca como ME_INTERESA o YA_JUGADO (usuario con sesión)",
        "tags": [
          "Biblioteca"
        ]
      }
    },
    "/biblioteca/{juegoId}": {
      "patch": {
        "operationId": "JuegoGuardadoController_cambiarEstado",
        "parameters": [
          {
            "name": "juegoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarEstadoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JuegoGuardado"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambiar el estado de un juego de mi biblioteca (usuario con sesión)",
        "tags": [
          "Biblioteca"
        ]
      },
      "delete": {
        "operationId": "JuegoGuardadoController_quitar",
        "parameters": [
          {
            "name": "juegoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Quitar un juego de mi biblioteca (usuario con sesión)",
        "tags": [
          "Biblioteca"
        ]
      }
    },
...
    "/recomendaciones/{busquedaId}/juegos/{juegoId}": {
      "patch": {
        "operationId": "RecomendacionController_calificar",
        "parameters": [
          {
            "name": "busquedaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "juegoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalificarRecomendacionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recomendacion"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Calificar de 1 a 5 un juego recomendado, con comentario opcional (el dueño o un ADMIN)",
        "tags": [
          "Recomendaciones"
        ]
      }
    },
...
  },
  "info": {
    "title": "DGame (MyGameSearcher) API",
    "description": "API REST que recomienda de 1 a 3 videojuegos según géneros, características y plataforma.\n\n**Cómo autenticarse:** hacer `POST /auth/login` con email y contraseña, copiar el `token` de la respuesta y pegarlo en el botón **Authorize** (arriba a la derecha). Desde ahí, todas las pruebas se mandan con `Authorization: Bearer <token>`.\n\n**Niveles de acceso:** público (sin token), usuario con sesión iniciada, y ADMIN. El resumen de cada ruta indica quién puede usarla.\n\n**Errores:** todas las respuestas de error tienen la forma `{ \"statusCode\", \"message\", \"error\" }`, con el mensaje en español. 400: datos inválidos · 401: sin sesión o token inválido · 403: sin permiso · 404: no existe · 409: conflicto (duplicado o en uso).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token que devuelve POST /auth/login"
      }
    },
    "schemas": {
...
      "Genero": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
      "ErrorRespuestaDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "description": "Mensaje en español. En los errores de validación (400) es una lista con un mensaje por campo",
            "oneOf": [
              {
                "type": "string",
                "example": "No se encontró el género con id 9999"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "El nombre es obligatorio"
                ]
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "Plataforma": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
...
      "Caracteristica": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
...
      "ClasificacionEdad": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
...
      "Juego": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "titulo": {
            "type": "string"
          },
          "anioLanzamiento": {
            "type": "number"
          },
          "descripcion": {
            "type": "string"
          },
          "imagenUrl": {
            "type": "string",
            "nullable": true
          },
          "clasificacionEdad": {
            "$ref": "#/components/schemas/ClasificacionEdad"
          },
          "plataformas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plataforma"
            }
          },
          "generos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genero"
            }
          },
          "caracteristicas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Caracteristica"
            }
          }
        },
        "required": [
          "id",
          "titulo",
          "anioLanzamiento",
          "descripcion",
          "imagenUrl",
          "clasificacionEdad",
          "plataformas",
          "generos",
          "caracteristicas"
        ]
      },
      "ActualizarJuegoDto": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 100
          },
          "anioLanzamiento": {
            "type": "number",
            "minimum": 1950,
            "maximum": 2026
          },
          "descripcion": {
            "type": "string",
            "maxLength": 2000
          },
          "imagenUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 500,
            "format": "uri"
          },
          "clasificacionEdadId": {
            "type": "number"
          },
          "plataformaIds": {
            "minItems": 1,
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "generoIds": {
            "minItems": 1,
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "caracteristicaIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "rol": {
            "enum": [
              "USUARIO",
              "ADMIN"
            ],
            "type": "string"
          },
          "fechaRegistro": {
            "format": "date-time",
            "type": "string"
          },
          "plataforma": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Plataforma"
              }
            ]
          }
        },
        "required": [
          "id",
          "nombre",
          "apellido",
          "email",
          "rol",
          "fechaRegistro",
          "plataforma"
        ]
      },
...
      "JuegoGuardado": {
        "type": "object",
        "properties": {
          "usuarioId": {
            "type": "number"
          },
          "juegoId": {
            "type": "number"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "juego": {
            "$ref": "#/components/schemas/Juego"
          },
          "estado": {
            "enum": [
              "ME_INTERESA",
              "YA_JUGADO"
            ],
            "type": "string"
          },
          "fecha": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "usuarioId",
          "juegoId",
          "usuario",
          "juego",
          "estado",
          "fecha"
        ]
      },
      "GuardarJuegoDto": {
        "type": "object",
        "properties": {
          "juegoId": {
            "type": "number"
          },
          "estado": {
            "type": "string",
            "enum": [
              "ME_INTERESA",
              "YA_JUGADO"
            ]
          }
        },
        "required": [
          "juegoId",
          "estado"
        ]
      },
      "CambiarEstadoDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "enum": [
              "ME_INTERESA",
              "YA_JUGADO"
            ]
          }
        },
        "required": [
          "estado"
        ]
      },
...
      "Busqueda": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "fechaBusqueda": {
            "format": "date-time",
            "type": "string"
          },
          "plataformas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plataforma"
            }
          },
          "generos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genero"
            }
          },
          "caracteristicas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Caracteristica"
            }
          },
          "recomendaciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recomendacion"
            }
          }
        },
        "required": [
          "id",
          "usuario",
          "fechaBusqueda",
          "plataformas",
          "generos",
          "caracteristicas",
          "recomendaciones"
        ]
      },
      "Recomendacion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "busqueda": {
            "$ref": "#/components/schemas/Busqueda"
          },
          "juego": {
            "$ref": "#/components/schemas/Juego"
          },
          "orden": {
            "type": "number"
          },
          "calificacion": {
            "type": "number",
            "nullable": true
          },
          "comentario": {
            "type": "string",
            "nullable": true
          },
          "fechaCalificacion": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "busqueda",
          "juego",
          "orden",
          "calificacion",
          "comentario",
          "fechaCalificacion"
        ]
      },
      "CalificarRecomendacionDto": {
        "type": "object",
        "properties": {
          "calificacion": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "comentario": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          }
        },
        "required": [
          "calificacion"
        ]
      },
...
    }
  }
}